requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.8
sentence-transformers==2.2.2
groq==0.4.0
gunicorn==20.1.0
